requirements.txt
#code checking library 
pylint

#Allows for the use of .env secret files. 
python-dotenv

#library containing technical analysis tools
pandas-ta

#library used to update postgreSQL
psycopg2

#pycoingecko library that hosts the coingecko public api.
pycoingecko